refactor(map): improve readability and fix errors
- added `setSelectedMarker(undefined)` to `onMapPress` function in `SueReportLocation` and `SueMapViewScreen` to remove callout text from the screen
- added `calculateBoundsToFitAllMarkers` helper to prevent the Map.tsx component from growing larger
- updated the name of the `calculateBounds` helper to `calculateBoundsToFitAllMarkers` for clarity
- added comment line to understand values in `bounds` array
- increased the `animationDuration` value from 0 to 1000 to solve the problem of the map updating itself according to the desired location after a certain period of time
- updated or deleted unnecessary codes

SVA-1365
